Refactor: Rename IDbContext to ISmarthomeContext
Renamed the `IDbContext` interface to `ISmarthomeContext` and its implementation `DbContext` to `SmarthomeContext`. Updated all ViewModel and test files to use the new interface and class names. Modified `MauiProgram.cs` to register `ISmarthomeContext` and `SmarthomeContext` in the dependency injection container. Removed the old `IDbContext` and `DbContext` files.
